+++ android/loginregistration/src/androidTest/java/com/ccorrads/ossp/loginregistration/ExampleInstrumentedTest.kt
@@ -1,0 +1,20 @@
+package com.ccorrads.ossp.loginregistration
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("com.ccorrads.ossp.loginregistration.test", appContext.packageName)
+    }
+}