+++ android/loginregistration/src/main/java/com/ccorrads/ossp/loginregistration/AuthMvp.kt
@@ -1,0 +1,16 @@
+package com.ccorrads.ossp.loginregistration
+
+import com.ccorrads.ossp.core.Mvp
+import com.ccorrads.ossp.loginregistration.views.ValidatableText
+
+class AuthMvp {
+
+    interface View : Mvp.View
+
+    interface Presenter : Mvp.Presenter {
+
+        fun validateViews(emailView: ValidatableText, pwView: ValidatableText): Boolean
+
+        fun onClickLogin(loginView: AuthMvp.View, user: String?, pw: String?)
+    }
+}