+++ android/loginregistration/src/main/java/com/ccorrads/ossp/loginregistration/injection/LoginRegisterModule.kt
@@ -1,0 +1,16 @@
+package com.ccorrads.ossp.loginregistration.injection
+
+import com.ccorrads.ossp.loginregistration.AuthMvp
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class LoginRegisterModule {
+
+    @Provides
+    @Singleton
+    fun presenter(presenter: LoginRegisterPresenter): AuthMvp.Presenter {
+        return presenter
+    }
+}